[#25081] Extract common file download actions
