+++ app/controllers/concerns/order_detail_file_download.rb
@@ -1,0 +1,13 @@
+module OrderDetailFileDownload
+  extend ActiveSupport::Concern
+
+  def sample_results
+    authorize! :sample_results, @order_detail
+    redirect_to @order_detail.stored_files.sample_result.find(params[:stored_file_id]).download_url
+  end
+
+  def template_results
+    authorize! :template_results, @order_detail
+    redirect_to @order_detail.stored_files.template_result.find(params[:stored_file_id]).download_url
+  end
+end
