+++ app/controllers/order_details_controller.rb
@@ -1,4 +1,6 @@
 class OrderDetailsController < ApplicationController
+  include OrderDetailFileDownload
+
   customer_tab  :all
 
   before_filter :authenticate_user!
@@ -99,18 +101,6 @@ class OrderDetailsController < ApplicationController
     end
     @order.invalidate!
     redirect_to(order_path(@order))
-  end
-
-  # GET /orders/:order_id/order_details/:id/sample_results/:stored_file_id
-  def sample_results
-    authorize! :sample_results, @order_detail
-    redirect_to @order_detail.stored_files.sample_result.find(params[:stored_file_id]).download_url
-  end
-
-  # GET /orders/:order_id/order_details/:id/template_results/:stored_file_id
-  def template_results
-    authorize! :template_results, @order_detail
-    redirect_to @order_detail.stored_files.template_result.find(params[:stored_file_id]).download_url
   end
 
   def set_active_tab
