+++ app/controllers/order_management/order_details_controller.rb
@@ -1,4 +1,6 @@
 class OrderManagement::OrderDetailsController < ApplicationController
+  include OrderDetailFileDownload
+
   load_resource :facility, :find_by => :url_name
   load_resource :order, :through => :facility
   load_resource :order_detail, :through => :order
@@ -47,18 +49,6 @@ class OrderManagement::OrderDetailsController < ApplicationController
   def files
     @files = @order_detail.stored_files.sample_result.order(:created_at)
     render :layout => false if request.xhr?
-  end
-
-  # GET /facilities/:facility_id/orders/:order_id/order_details/:id/sample_results/:stored_file_id
-  def sample_results
-    authorize! :sample_results, @order_detail
-    redirect_to @order_detail.stored_files.sample_result.find(params[:stored_file_id]).download_url
-  end
-
-  # GET /facilities/:facility_id/orders/:order_id/order_details/:id/template_results/:stored_file_id
-  def template_results
-    authorize! :template_results, @order_detail
-    redirect_to @order_detail.stored_files.template_result.find(params[:stored_file_id]).download_url
   end
 
   # POST /facilities/:facility_id/orders/:order_id/order_details/:id/remove_from_journal
